requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.10
botocore==1.38.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
distro==1.9.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langsmith==0.3.42
matplotlib==3.10.1
numpy==2.2.5
openai==1.77.0
orjson==3.10.18
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pinecone==6.0.0
pinecone-plugin-interface==0.0.7
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
qstash==2.0.5
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
upstash-workflow==0.1.2
urllib3==2.4.0
uvicorn==0.34.2
zstandard==0.23.0
